pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Jehad Abu Awwad <jehadabuawwad@outlook.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
